test(domain): 기존 Lotto에서 진행하던 당첨 번호 확인 테스트 WinningLottoTest에서 테스트 진행
- 테스트 메서드를 옮기에 따른 로직 변경
