+++ src/test/java/lotto/domain/LottoTest.java
@@ -25,80 +25,4 @@ class LottoTest {
         assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 5)))
                 .isInstanceOf(IllegalArgumentException.class);
     }
-
-    @DisplayName("당첨 번호와 일치하는 개수 반환")
-    @MethodSource("provideMatchCases")
-    @ParameterizedTest
-    void 당첨_번호와_일치하는_개수_반환(
-            List<Integer> myNumbers,
-            List<Integer> winningNumber,
-            int expectedCount
-    ) {
-        //given
-        Lotto input = new Lotto(myNumbers);
-
-        //when
-        int result = input.countMatches(winningNumber);
-
-        //then
-        assertThat(result).isEqualTo(expectedCount);
-    }
-
-    private static Stream<Arguments> provideMatchCases() {
-        return Stream.of(
-                Arguments.of(
-                        List.of(1, 2, 3, 4, 5, 6),
-                        List.of(1, 2, 3, 4, 5, 6),
-                        6
-                ),
-                Arguments.of(
-                        List.of(1, 2, 3, 4, 5, 6),
-                        List.of(1, 2, 3, 4, 5, 7),
-                        5
-                ),
-                Arguments.of(
-                        List.of(1, 2, 3, 4, 5, 6),
-                        List.of(1, 2, 3, 4, 8, 7),
-                        4
-                ),
-                Arguments.of(
-                        List.of(1, 2, 3, 4, 5, 6),
-                        List.of(1, 2, 3, 9, 8, 7),
-                        3
-                ),
-                Arguments.of(
-                        List.of(1, 2, 3, 4, 5, 6),
-                        List.of(1, 2, 10, 9, 8, 7),
-                        2
-                ),
-                Arguments.of(
-                        List.of(1, 2, 3, 4, 5, 6),
-                        List.of(1, 11, 10, 9, 8, 7),
-                        1
-                ),
-                Arguments.of(
-                        List.of(1, 2, 3, 4, 5, 6),
-                        List.of(12, 11, 10, 9, 8, 7),
-                        0
-                ),
-                Arguments.of(
-                        List.of(1, 2, 3, 4, 5, 6),
-                        List.of(6, 5, 4, 3, 2, 1),
-                        6
-                )
-        );
-    }
-
-    @DisplayName("보너스 번호가 있는지 확인")
-    @Test
-    void 보너스_번호가_있는지_확인() {
-        //given
-        Lotto myLotto = new Lotto(List.of(1, 2, 3, 4, 5, 6));
-        int bonusNumber = 6;
-
-        //when
-        boolean result = myLotto.hasBonusNumber(bonusNumber);
-
-        assertThat(result).isTrue();
-    }
 }
